backend/requirements.txt
google-cloud-aiplatform
streamlit 
speechrecognition 
pyttsx3 
vertexai
gTTS
paho-mqtt
fastapi
uvicorn